- Trying to get rid of the distinction between "manager" and "entity system". They're all EntityObservers now. - Now both 'init' and 'dispose' methods are available to all entity observers. Also added a World instance and 'process' method for all of them. - Had to refactor MapperImplementor and World instances a bit to get it to compile.

+++ src/com/artemis/PooledEntityManager.java
@@ -44,7 +44,7 @@ public final class PooledEntityManager extends EntityManager
 		entityStore.addAll( eArray, eSize - (newSize - poolSize) );
 	}
 
-	public long getTotalPooled ()
+	public int getTotalPooled ()
 	{
 		return entityStore.size();
 	}
